+++ examples/react-template/screens/Divider.tsx
@@ -1,5 +1,5 @@
 import { Row, Rows, Spacer, SpacerSize } from '@trilogy-ds/react'
-import { Box, BoxContent, Divider, IconName, Section, Text } from '@trilogy-ds/react/lib/components'
+import { Box, BoxContent, Divider, IconName, Section, Text } from '@trilogy-ds/react/components'
 import { GapSize } from '@trilogy-ds/react/lib/components/columns/ColumnsTypes'
 import * as React from 'react'