+++ examples/react-template/screens/Textarea.tsx
@@ -1,5 +1,5 @@
-import { IconName } from '@trilogy-ds/react'
-import { Divider, InputStatus, Section, Textarea } from '@trilogy-ds/react/lib/components'
+import { IconName, InputStatus } from '@trilogy-ds/react'
+import { Divider, Section, Textarea } from '@trilogy-ds/react/components'
 import * as React from 'react'
 
 export const TextareaScreen = (): JSX.Element => {
@@ -13,7 +13,6 @@ export const TextareaScreen = (): JSX.Element => {
         rows={10}
         iconNameLeft={IconName.CHECK}
         iconNameRight='tri-exclamation-circle'
-        dynamicPlaceholder={false}
         status={InputStatus.ERROR}
       />
 
@@ -21,23 +20,29 @@ export const TextareaScreen = (): JSX.Element => {
 
       <Textarea
         label='Textarea label not dynamic without sample'
+        sample='Textarea sample'
         help='Search helper textarea'
         required
         iconNameLeft={IconName.CHECK}
-        dynamicPlaceholder={false}
       />
 
       <Divider />
 
-      <Textarea disabled placeholder='placeholder' label='Dynamic label' />
+      <Textarea
+        disabled
+        placeholder='placeholder'
+        label='Dynamic label'
+        iconNameLeft='tri-alert'
+        iconNameRight='tri-alert'
+      />
 
       <Divider />
 
-      <Textarea placeholder='placeholder' label='No Dynamic label' dynamicPlaceholder={false} />
+      <Textarea placeholder='placeholder' label='No Dynamic label' />
 
       <Divider />
 
-      <Textarea placeholder='placeholder' label='With Max Lenght ' dynamicPlaceholder={false} maxLength={150} />
+      <Textarea placeholder='placeholder' label='With Max Lenght ' maxLength={150} />
       <Divider />
       <Textarea placeholder='Disabled' label='Label' disabled />